Crate nri

This crate contains all the software used to run the sensing rig. It's designed to run on the Intel NUC in the rig operator's backpack. There are driver wrappers for each sensor, a supervisor that runs each of those in its own thread, plus CLI and web interfaces to control everything.

Reexports

extern crate "std" as std;

extern crate log;

extern crate env_logger;

use std::prelude::v1::*;

use std::io;

use std::io::{Write, BufRead};

use std::ascii::AsciiExt;

use std::ptr;

use std::thread;

use std::sync::mpsc::{Receiver, TryRecvError};

use std::process::Command;

use mpmc::MultiSender;

use comms::Cmd;

use web::Web;

use structure::Structure;

use bluefox::Bluefox;

use optoforce::Optoforce;

Modules

bluefox

comms
COMMS DOCS HERE

mpmc
thanks "panicbit" on irc.mozilla.org #rust

optoforce

structure

web
Web interface to view and control running services

Structs

Service
Service descriptor

Functions

main
Main function that does everything