Struct nri::web::Service

struct Service {
    name: String,
}

Service descriptor

Unlike the one in main.rs, this descriptor only needs to contain things that are useful for display in the interface. However, they should probably be unified (TODO). The "web descriptor" could be just a subfield of the main.rs service descriptor, and then those could get passed in here (somehow).

Fields

name

Methods

impl Service

fn new(s: &str) -> Service

Create a new service descriptor with the given name