Crate nri

This crate contains all the software used to run the sensing rig. It's designed to run on the Intel NUC in the rig operator's backpack. There are driver wrappers for each sensor, a supervisor that runs each of those in its own thread, plus CLI and web interfaces to control everything.

HOWTO

How to generate this documentation

Run cargo doc in the NRI repo:

nri$ cargo doc

2. Now we need the rustdoc command for this crate.

nri$ touch src/main.rs
nri$ cargo doc -v | grep rustdoc | awk "-FRunning " '{print substr($NF, 2, length($NF)-2)}' > cargo-doc-command

3. Edit <tt>cargo-doc-command</tt> and add <tt>--no-defaults --passes "collapse-docs" --passes "unindent-comments"</tt> after <tt>src/main.rs</tt>. Then run it.

nri$ source cargo-doc-command

4. Copy the docs to the Github Pages repo, commit, and push.

nri$ rsync -a target/doc ../haptics-nri.github.io
nri$ cd ../haptics-nri.github.io
haptics-nri.github.io$ git add doc
haptics-nri.github.io$ git commit -m "cargo doc"
haptics-nri.github.io$ git push

5. The docs are now live (after 30s or so) at http://haptics-nri.github.io/doc/nri.

Reexports

extern crate "std" as std;

extern crate log;

extern crate env_logger;

use std::prelude::v1::*;

use std::io;

use std::io::{Write, BufRead};

use std::ascii::AsciiExt;

use std::ptr;

use std::thread;

use std::sync::mpsc::{Receiver, TryRecvError};

use std::process::Command;

use mpmc::MultiSender;

use comms::Cmd;

use web::Web;

use structure::Structure;

use bluefox::Bluefox;

use optoforce::Optoforce;

Modules

bluefox

comms
COMMS DOCS HERE

mpmc
thanks "panicbit" on irc.mozilla.org #rust

optoforce

structure

web
Web interface to view and control running services

Structs

Service
Service descriptor

Functions

main
Main function that does everything